Refactor calendars: extract from `utils.py` to `calendars.py`. (#969)
