+++ holidays/countries/albania.py
@@ -14,9 +14,9 @@ from datetime import timedelta as td
 
 from dateutil.easter import EASTER_ORTHODOX, easter
 
+from holidays.calendars import _islamic_to_gre
 from holidays.constants import JAN, MAR, MAY, SEP, OCT, NOV, DEC
 from holidays.holiday_base import HolidayBase
-from holidays.utils import _islamic_to_gre
 
 
 class Albania(HolidayBase):
