+++ holidays/countries/kyrgyzstan.py
@@ -12,9 +12,9 @@
 from datetime import date
 from datetime import timedelta as td
 
+from holidays.calendars import _islamic_to_gre
 from holidays.constants import JAN, FEB, MAR, APR, MAY, AUG, NOV, DEC
 from holidays.holiday_base import HolidayBase
-from holidays.utils import _islamic_to_gre
 
 
 class Kyrgyzstan(HolidayBase):
