+++ holidays/countries/saudi_arabia.py
@@ -12,9 +12,9 @@
 from datetime import date
 from datetime import timedelta as td
 
+from holidays.calendars import _islamic_to_gre
 from holidays.constants import FEB, SEP, NOV, THU, FRI, SAT
 from holidays.holiday_base import HolidayBase
-from holidays.utils import _islamic_to_gre
 
 
 class SaudiArabia(HolidayBase):
