+++ holidays/countries/spain.py
@@ -14,10 +14,10 @@ from datetime import timedelta as td
 
 from dateutil.easter import easter
 
+from holidays.calendars import _islamic_to_gre
 from holidays.constants import JAN, FEB, MAR, APR, MAY, JUN, JUL, AUG, SEP
 from holidays.constants import OCT, NOV, DEC
 from holidays.holiday_base import HolidayBase
-from holidays.utils import _islamic_to_gre
 
 
 class Spain(HolidayBase):
