+++ holidays/countries/united_arab_emirates.py
@@ -12,10 +12,10 @@
 from datetime import date
 from datetime import timedelta as td
 
+from holidays.calendars import _islamic_to_gre
 from holidays.constants import JAN, APR, MAY, JUN, JUL, AUG, SEP, NOV, DEC
 from holidays.constants import FRI, SAT
 from holidays.holiday_base import HolidayBase
-from holidays.utils import _islamic_to_gre
 
 
 class UnitedArabEmirates(HolidayBase):
